+++ packages/provider/lib/src/change_notifier_provider.dart
@@ -21,10 +21,10 @@ import 'proxy_provider.dart';
 /// which explains in further details why using the `.value` constructor to
 /// create values is undesired.
 ///
-/// - DO create a new [ChangeNotifier] inside `builder`.
+/// - DO create a new [ChangeNotifier] inside `create`.
 /// ```dart
 /// ChangeNotifierProvider(
-///   builder: (_) => new MyChangeNotifier(),
+///   create: (_) => new MyChangeNotifier(),
 ///   child: ...
 /// )
 /// ```
@@ -46,7 +46,7 @@ import 'proxy_provider.dart';
 /// int count;
 ///
 /// ChangeNotifierProvider(
-///   builder: (_) => new MyChangeNotifier(count),
+///   create: (_) => new MyChangeNotifier(count),
 ///   child: ...
 /// )
 /// ```
@@ -80,7 +80,7 @@ import 'proxy_provider.dart';
 /// MyChangeNotifier variable;
 ///
 /// ChangeNotifierProvider(
-///   builder: (_) => variable,
+///   create: (_) => variable,
 ///   child: ...
 /// )
 /// ```
@@ -97,15 +97,24 @@ class ChangeNotifierProvider<T extends ChangeNotifier>
   static void _disposer(BuildContext context, ChangeNotifier notifier) =>
       notifier?.dispose();
 
-  /// Creates a [ChangeNotifier] using `builder` and automatically
+  /// Creates a [ChangeNotifier] using `create` and automatically
   /// dispose it when [ChangeNotifierProvider] is removed from the widget tree.
   ///
-  /// `builder` must not be `null`.
+  /// `create` must not be `null`.
   ChangeNotifierProvider({
     Key key,
-    @required ValueBuilder<T> builder,
+    @required ValueBuilder<T> create,
+    @Deprecated('will be removed in 4.0.0, use create instead')
+        ValueBuilder<T> builder,
     Widget child,
-  }) : super(key: key, builder: builder, dispose: _disposer, child: child);
+  }) : super(
+          key: key,
+          create:
+              // ignore: deprecated_member_use_from_same_package
+              create ?? builder,
+          dispose: _disposer,
+          child: child,
+        );
 
   /// Provides an existing [ChangeNotifier].
   ChangeNotifierProvider.value({
@@ -124,7 +133,7 @@ class ChangeNotifierProvider<T extends ChangeNotifier>
 ///
 /// ```dart
 /// ChangeNotifierProvider(
-///   builder: (context) {
+///   create: (context) {
 ///     return MyChangeNotifier(
 ///       foo: Provider.of<Foo>(context, listen: false),
 ///     );
@@ -143,8 +152,8 @@ class ChangeNotifierProvider<T extends ChangeNotifier>
 ///
 /// ```dart
 /// ChangeNotifierProxyProvider<Foo, MyChangeNotifier>(
-///   initialBuilder: (_) => MyChangeNotifier(),
-///   builder: (_, foo, myNotifier) => myNotifier
+///   create: (_) => MyChangeNotifier(),
+///   update: (_, foo, myNotifier) => myNotifier
 ///     ..foo = foo,
 ///   child: ...
 /// );
@@ -170,7 +179,7 @@ class ChangeNotifierProvider<T extends ChangeNotifier>
 /// }
 /// ```
 ///
-/// - DON'T create the [ChangeNotifier] inside `builder` directly.
+/// - DON'T create the [ChangeNotifier] inside `update` directly.
 ///
 ///   This will cause your state to be lost when one of the values used updates.
 ///   It will also cause uncesserary overhead because it will dispose the
@@ -182,7 +191,7 @@ class ChangeNotifierProvider<T extends ChangeNotifier>
 /// ```dart
 /// ChangeNotifierProxyProvider<Foo, MyChangeNotifier>(
 ///   // may cause the state to be destroyed unvoluntarily
-///   builder: (_, foo, myNotifier) => MyChangeNotifier(foo: foo),
+///   update: (_, foo, myNotifier) => MyChangeNotifier(foo: foo),
 ///   child: ...
 /// );
 /// ```
@@ -198,13 +207,19 @@ class ChangeNotifierProxyProvider<T, R extends ChangeNotifier>
   /// Initializes [key] for subclasses.
   ChangeNotifierProxyProvider({
     Key key,
-    @required ValueBuilder<R> initialBuilder,
-    @required ProxyProviderBuilder<T, R> builder,
+    @required ValueBuilder<R> create,
+    @required ProxyProviderBuilder<T, R> update,
+    @Deprecated('will be removed in 4.0.0, use create instead')
+        ValueBuilder<R> initialBuilder,
+    @Deprecated('will be removed in 4.0.0, use update instead')
+        ProxyProviderBuilder<T, R> builder,
     Widget child,
   }) : super(
           key: key,
-          initialBuilder: initialBuilder,
-          builder: builder,
+          // ignore: deprecated_member_use_from_same_package
+          create: create ?? initialBuilder,
+          // ignore: deprecated_member_use_from_same_package
+          update: update ?? builder,
           dispose: ChangeNotifierProvider._disposer,
           child: child,
         );
@@ -216,13 +231,19 @@ class ChangeNotifierProxyProvider2<T, T2, R extends ChangeNotifier>
   /// Initializes [key] for subclasses.
   ChangeNotifierProxyProvider2({
     Key key,
-    @required ValueBuilder<R> initialBuilder,
-    @required ProxyProviderBuilder2<T, T2, R> builder,
+    @required ValueBuilder<R> create,
+    @required ProxyProviderBuilder2<T, T2, R> update,
+    @Deprecated('will be removed in 4.0.0, use create instead')
+        ValueBuilder<R> initialBuilder,
+    @Deprecated('will be removed in 4.0.0, use update instead')
+        ProxyProviderBuilder2<T, T2, R> builder,
     Widget child,
   }) : super(
           key: key,
-          initialBuilder: initialBuilder,
-          builder: builder,
+          // ignore: deprecated_member_use_from_same_package
+          create: create ?? initialBuilder,
+          // ignore: deprecated_member_use_from_same_package
+          update: update ?? builder,
           dispose: ChangeNotifierProvider._disposer,
           child: child,
         );
@@ -234,13 +255,19 @@ class ChangeNotifierProxyProvider3<T, T2, T3, R extends ChangeNotifier>
   /// Initializes [key] for subclasses.
   ChangeNotifierProxyProvider3({
     Key key,
-    @required ValueBuilder<R> initialBuilder,
-    @required ProxyProviderBuilder3<T, T2, T3, R> builder,
+    @required ValueBuilder<R> create,
+    @required ProxyProviderBuilder3<T, T2, T3, R> update,
+    @Deprecated('will be removed in 4.0.0, use create instead')
+        ValueBuilder<R> initialBuilder,
+    @Deprecated('will be removed in 4.0.0, use update instead')
+        ProxyProviderBuilder3<T, T2, T3, R> builder,
     Widget child,
   }) : super(
           key: key,
-          initialBuilder: initialBuilder,
-          builder: builder,
+          // ignore: deprecated_member_use_from_same_package
+          create: create ?? initialBuilder,
+          // ignore: deprecated_member_use_from_same_package
+          update: update ?? builder,
           dispose: ChangeNotifierProvider._disposer,
           child: child,
         );
@@ -252,13 +279,19 @@ class ChangeNotifierProxyProvider4<T, T2, T3, T4, R extends ChangeNotifier>
   /// Initializes [key] for subclasses.
   ChangeNotifierProxyProvider4({
     Key key,
-    @required ValueBuilder<R> initialBuilder,
-    @required ProxyProviderBuilder4<T, T2, T3, T4, R> builder,
+    @required ValueBuilder<R> create,
+    @required ProxyProviderBuilder4<T, T2, T3, T4, R> update,
+    @Deprecated('will be removed in 4.0.0, use create instead')
+        ValueBuilder<R> initialBuilder,
+    @Deprecated('will be removed in 4.0.0, use update instead')
+        ProxyProviderBuilder4<T, T2, T3, T4, R> builder,
     Widget child,
   }) : super(
           key: key,
-          initialBuilder: initialBuilder,
-          builder: builder,
+          // ignore: deprecated_member_use_from_same_package
+          create: create ?? initialBuilder,
+          // ignore: deprecated_member_use_from_same_package
+          update: update ?? builder,
           dispose: ChangeNotifierProvider._disposer,
           child: child,
         );
@@ -271,13 +304,19 @@ class ChangeNotifierProxyProvider5<T, T2, T3, T4, T5, R extends ChangeNotifier>
   /// Initializes [key] for subclasses.
   ChangeNotifierProxyProvider5({
     Key key,
-    @required ValueBuilder<R> initialBuilder,
-    @required ProxyProviderBuilder5<T, T2, T3, T4, T5, R> builder,
+    @required ValueBuilder<R> create,
+    @required ProxyProviderBuilder5<T, T2, T3, T4, T5, R> update,
+    @Deprecated('will be removed in 4.0.0, use create instead')
+        ValueBuilder<R> initialBuilder,
+    @Deprecated('will be removed in 4.0.0, use update instead')
+        ProxyProviderBuilder5<T, T2, T3, T4, T5, R> builder,
     Widget child,
   }) : super(
           key: key,
-          initialBuilder: initialBuilder,
-          builder: builder,
+          // ignore: deprecated_member_use_from_same_package
+          create: create ?? initialBuilder,
+          // ignore: deprecated_member_use_from_same_package
+          update: update ?? builder,
           dispose: ChangeNotifierProvider._disposer,
           child: child,
         );
@@ -290,13 +329,19 @@ class ChangeNotifierProxyProvider6<T, T2, T3, T4, T5, T6,
   /// Initializes [key] for subclasses.
   ChangeNotifierProxyProvider6({
     Key key,
-    @required ValueBuilder<R> initialBuilder,
-    @required ProxyProviderBuilder6<T, T2, T3, T4, T5, T6, R> builder,
+    @required ValueBuilder<R> create,
+    @required ProxyProviderBuilder6<T, T2, T3, T4, T5, T6, R> update,
+    @Deprecated('will be removed in 4.0.0, use create instead')
+        ValueBuilder<R> initialBuilder,
+    @Deprecated('will be removed in 4.0.0, use update instead')
+        ProxyProviderBuilder6<T, T2, T3, T4, T5, T6, R> builder,
     Widget child,
   }) : super(
           key: key,
-          initialBuilder: initialBuilder,
-          builder: builder,
+          // ignore: deprecated_member_use_from_same_package
+          create: create ?? initialBuilder,
+          // ignore: deprecated_member_use_from_same_package
+          update: update ?? builder,
           dispose: ChangeNotifierProvider._disposer,
           child: child,
         );