+++ packages/provider/test/listenable_proxy_provider_test.dart
@@ -28,11 +28,11 @@ void main() {
   );
 
   group('ListenableProxyProvider', () {
-    test('throws if builder is missing', () {
+    test('throws if update is missing', () {
       expect(
         () => ListenableProxyProvider<A, _ListenableCombined>(
-          initialBuilder: null,
-          builder: null,
+          create: null,
+          update: null,
         ),
         throwsAssertionError,
       );
@@ -46,8 +46,8 @@ void main() {
           providers: [
             Provider.value(value: 0),
             ListenableProxyProvider<int, ValueNotifier<int>>(
-              initialBuilder: null,
-              builder: (_, __, ___) => notifier,
+              create: null,
+              update: (_, __, ___) => notifier,
               dispose: (_, __) {},
             )
           ],
@@ -64,8 +64,8 @@ void main() {
           providers: [
             Provider.value(value: 0),
             ListenableProxyProvider<int, ValueNotifier<int>>(
-              initialBuilder: null,
-              builder: (_, __, ___) => ValueNotifier(0),
+              create: null,
+              update: (_, __, ___) => ValueNotifier(0),
               dispose: (_, __) {},
             )
           ],
@@ -77,8 +77,8 @@ void main() {
           providers: [
             Provider.value(value: 1),
             ListenableProxyProvider<int, ValueNotifier<int>>(
-              initialBuilder: null,
-              builder: (_, __, ___) => notifier,
+              create: null,
+              update: (_, __, ___) => notifier,
               dispose: (_, __) {},
             )
           ],
@@ -96,8 +96,8 @@ void main() {
           providers: [
             Provider.value(value: 0),
             ListenableProxyProvider<int, ValueNotifier<int>>(
-              initialBuilder: (_) => notifier,
-              builder: (_, count, value) => value..value = count,
+              create: (_) => notifier,
+              update: (_, count, value) => value..value = count,
             )
           ],
           child: Consumer<ValueNotifier<int>>(builder: (_, value, __) {
@@ -119,7 +119,7 @@ void main() {
       expect(find.text('0'), findsNothing);
     });
     testWidgets(
-      'builder returning a new Listenable disposes the previously created value'
+      'update returning a new Listenable disposes the previously created value'
       ' and update dependents',
       (tester) async {
         final builder = MockConsumerBuilder<MockNotifier>();
@@ -134,8 +134,8 @@ void main() {
             providers: [
               Provider.value(value: 0),
               ListenableProxyProvider<int, MockNotifier>(
-                initialBuilder: null,
-                builder: (_, count, value) => notifier,
+                create: null,
+                update: (_, count, value) => notifier,
                 dispose: dispose,
               )
             ],
@@ -153,8 +153,8 @@ void main() {
             providers: [
               Provider.value(value: 1),
               ListenableProxyProvider<int, MockNotifier>(
-                initialBuilder: null,
-                builder: (_, count, value) => notifier2,
+                create: null,
+                update: (_, count, value) => notifier2,
                 dispose: dispose2,
               )
             ],
@@ -185,8 +185,8 @@ void main() {
             Provider.value(value: 0),
             ListenableProxyProvider<int, ValueNotifier<int>>(
               key: key,
-              initialBuilder: (_) => notifier,
-              builder: (_, count, value) => value..value = count,
+              create: (_) => notifier,
+              update: (_, count, value) => value..value = count,
               dispose: dispose,
             )
           ],
@@ -218,8 +218,8 @@ void main() {
             Provider.value(value: e),
             Provider.value(value: f),
             ListenableProxyProvider2<A, B, _ListenableCombined>(
-              initialBuilder: (_) => _ListenableCombined(null, null, null),
-              builder: (context, a, b, previous) =>
+              create: (_) => _ListenableCombined(null, null, null),
+              update: (context, a, b, previous) =>
                   _ListenableCombined(context, previous, a, b),
             )
           ],
@@ -247,8 +247,8 @@ void main() {
             Provider.value(value: e),
             Provider.value(value: f),
             ListenableProxyProvider3<A, B, C, _ListenableCombined>(
-              initialBuilder: (_) => _ListenableCombined(null, null, null),
-              builder: (context, a, b, c, previous) =>
+              create: (_) => _ListenableCombined(null, null, null),
+              update: (context, a, b, c, previous) =>
                   _ListenableCombined(context, previous, a, b, c),
             )
           ],
@@ -276,8 +276,8 @@ void main() {
             Provider.value(value: e),
             Provider.value(value: f),
             ListenableProxyProvider4<A, B, C, D, _ListenableCombined>(
-              initialBuilder: (_) => _ListenableCombined(null, null, null),
-              builder: (context, a, b, c, d, previous) =>
+              create: (_) => _ListenableCombined(null, null, null),
+              update: (context, a, b, c, d, previous) =>
                   _ListenableCombined(context, previous, a, b, c, d),
             )
           ],
@@ -305,8 +305,8 @@ void main() {
             Provider.value(value: e),
             Provider.value(value: f),
             ListenableProxyProvider5<A, B, C, D, E, _ListenableCombined>(
-              initialBuilder: (_) => _ListenableCombined(null, null, null),
-              builder: (context, a, b, c, d, e, previous) =>
+              create: (_) => _ListenableCombined(null, null, null),
+              update: (context, a, b, c, d, e, previous) =>
                   _ListenableCombined(context, previous, a, b, c, d, e),
             )
           ],
@@ -334,8 +334,8 @@ void main() {
             Provider.value(value: e),
             Provider.value(value: f),
             ListenableProxyProvider6<A, B, C, D, E, F, _ListenableCombined>(
-              initialBuilder: (_) => _ListenableCombined(null, null, null),
-              builder: (context, a, b, c, d, e, f, previous) =>
+              create: (_) => _ListenableCombined(null, null, null),
+              update: (context, a, b, c, d, e, f, previous) =>
                   _ListenableCombined(context, previous, a, b, c, d, e, f),
             )
           ],